refactor(payroll): PayrollCalculator 681→507 lignes — extraction de la régularisation dans PayrollRegularizationService (suite #5591, slice 3)
Suite du découpage du god-object : le concern « run de régularisation »
(calculateRegularizationRun + createDeltaLines — différentiel corrigé−original
par employé et par ligne, delta nul = pas de bulletin) est extrait dans
PayrollRegularizationService, comportement STRICTEMENT IDENTIQUE (copie).

- PayrollCalculator : orchestration (calculateRun, executeCalculateRun,
  calculateSlip) + façades ; délégation via la propriété regularization
  (instance partagée du PaySlipValueCalculator).
- PayrollRegularizationService : PaySlipValueCalculator injecté, DB::transaction
  interne conservée, gardes original verrouillé conservées.
- Filet : 892 tests verts (448 golden multi-pays, régularisation, delta,
  simulateurs) — 0 régression vs main (1 risky pré-existant sans assertion).
